kernel-3.x.x: add support SoC MT7621 and MT7628, giga-switch MT7530, updates for MT7620 rev.0205, remove RT2880 code, refactoring raeth code.

+++ trunk/linux-3.0.x/drivers/net/raeth/ra_rfrw.c
@@ -11,7 +11,7 @@
 
 int rw_rf_reg(int write, int reg, int *data)
 {
-	unsigned long rfcsr, i = 0;
+	u32 rfcsr, i = 0;
 
 	while (1) {
 		rfcsr = sysRegRead(RF_CSR_CFG);
